+++ assets/javascripts/kitten/constants/colors-config.js
@@ -26,4 +26,7 @@ export default {
   error: '#ff0046', // Red
   error2: '#ffe5ec', // Light Red
   error3: '#ffb2c7', // Semi Light Red
+
+  warning: '#8a6d3b',
+  warning2: '#fcf8e3',
 }